added error checking to scripts

+++ Software/Bash/killswitch-uninstall.sh
@@ -13,7 +13,7 @@
 #-------------------------------------------------------------------------------
 # constants
 
-VERSION_NUMBER="0.5.3-b1"
+VERSION_NUMBER="0.5.3-b2"
 SETTINGS_DIR="/home/${SUDO_USER}/.killswitch"
 
 #-------------------------------------------------------------------------------
@@ -51,23 +51,53 @@ echo ""
 # remove boot service script
 echo -n "Removing killswitch-boot.service from /lib/systemd/system/... "
 systemctl disable killswitch-boot.service
+if [ "$?" != "0" ]; then
+    echo "Failed"
+    echo "Aborting uninstall"
+    exit 1
+fi
 rm /lib/systemd/system/killswitch-boot.service
+if [ "$?" != "0" ]; then
+    echo "Failed"
+    echo "Aborting uninstall"
+    exit 1
+fi
 echo "Done"
 
 # remove monitor script
 echo -n "Removing killswitch-boot.py from /usr/local/bin/... "
 rm /usr/local/bin/killswitch-boot.py
+if [ "$?" != "0" ]; then
+    echo "Failed"
+    echo "Aborting uninstall"
+    exit 1
+fi
 echo "Done"
 
 # remove shutodwn service script
 echo -n "Removing killswitch-shutdown.service from /lib/systemd/system/... "
 systemctl disable killswitch-shutdown.service
+if [ "$?" != "0" ]; then
+    echo "Failed"
+    echo "Aborting uninstall"
+    exit 1
+fi
 rm /lib/systemd/system/killswitch-shutdown.service
+if [ "$?" != "0" ]; then
+    echo "Failed"
+    echo "Aborting uninstall"
+    exit 1
+fi
 echo "Done"
 
 # remove shutdown script
 echo -n "Removing killswitch-shutdown.py from /usr/local/bin... "
 rm /usr/local/bin/killswitch-shutdown.py
+if [ "$?" != "0" ]; then
+    echo "Failed"
+    echo "Aborting uninstall"
+    exit 1
+fi
 echo "Done"
 
 # remove settings gui script
@@ -78,17 +108,32 @@ echo "Done"
 # remove settings gui script
 echo -n "Removing kilswitch-settings.py from /usr/local/bin/... "
 rm /usr/local/bin/killswitch-settings.py
+if [ "$?" != "0" ]; then
+    echo "Failed"
+    echo "Aborting uninstall"
+    exit 1
+fi
 echo "Done"
 
 # remove uninstaller script
 echo -n "Removing kilswitch-uninstaller.sh from /usr/local/bin/... "
 rm /usr/local/bin/killswitch-uninstall.sh
+if [ "$?" != "0" ]; then
+    echo "Failed"
+    echo "Aborting uninstall"
+    exit 1
+fi
 echo "Done"
 
 # remove settings storage dir
 if [ -d "$SETTINGS_DIR" ]; then
     echo -n "Removing ${SETTINGS_DIR}... "
     rm -r "${SETTINGS_DIR}"
+    if [ "$?" != "0" ]; then
+        echo "Failed"
+        echo "Aborting uninstall"
+        exit 1
+    fi
     echo "Done"
 fi
 
@@ -100,6 +145,11 @@ echo ""
 if [ -d "/home/${SUDO_USER}/RetroPie" ]; then
 	echo -n "Removing RetroPie port... "
 	rm "/home/${SUDO_USER}/RetroPie/roms/ports/KillSwitch"
+    if [ "$?" != "0" ]; then
+        echo "Failed"
+        echo "Aborting uninstall"
+        exit 1
+    fi
 	echo "Done"
     echo ""
 fi
